fix: log out user if session has expired
Before this fix, the app would try to send the get requests to categories, tags, and notes routes without knowing that the session has expired. This would result in the app crashing (server), since those routes rely on req.session object which is removed by the express-session module automatically on session expiry.