feat: move add item button to menu bar in mobile layout
The add item (note/category/tag) button would be floating in the bottom right corner for both the mobile view with navigation menu bar up top and for the desktop view.

After this commit, the button will be placed in the nav menu bar for mobile devices and will stay floating for devices that use larger screens.